[Game Actions] Using hpp for headers.
