+++ tools/game-actions/game-actions.vcxproj
@@ -159,7 +159,7 @@
     <ClCompile Include="src\Main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="src\GameActions.h" />
+    <ClInclude Include="src\GameActions.hpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\libraries\core\core.vcxproj">
